requirements.txt
-e git+https://github.com/e-dream-ai/python-api.git@main#egg=edream_sdk
requests>=2.31.0
python-dotenv>=1.0.0
